pyproject.toml
[project]
name = "TelegramBotSupport"
version = "0.1.0"
description = "Telegram bot created for tech support functioning via topics"
authors = [
    {name = "Fraybyl",email = "fraybyl@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram (>=3.19.0,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

